Refatorando solução

+++ Lista_de_entregas/Models/IEntregas.cs
@@ -12,8 +12,8 @@ namespace Lista_de_entregas.Models
         string Endereco { get; set; }
         string Cidade { get; set; }
         estados Estados { get; set; }
-        float Frete { get; set; }
-        float Peso { get; set; }
+        double Frete { get; set; }
+        double Peso { get; set; }
         DateTime DataEntrega { get; set; }
     }
 }
